python-client/pyproject.toml
[project]
name = "hc-hermes"
version = "0.1.0"
description = "Python client library for HashiCorp Hermes document management system"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MPL-2.0" }
authors = [
    { name = "HashiCorp Labs", email = "labs@hashicorp.com" }
]
keywords = ["hermes", "hashicorp", "document-management", "api-client"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL-2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
cli = [
    "click>=8.1.7",
    "rich>=13.9.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-httpx>=0.32.0",
    "mypy>=1.13.0",
    "ruff>=0.7.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
hermes = "hc_hermes.cli:main"

[project.urls]
Homepage = "https://github.com/hashicorp/hermes"
Repository = "https://github.com/hashicorp/hermes"
Documentation = "https://github.com/hashicorp/hermes/tree/main/python-client"
Issues = "https://github.com/hashicorp/hermes/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hc_hermes"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "TRY",    # tryceratops
    "RUF",    # ruff-specific rules
]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
    "D",       # pydocstyle (we'll add this later)
    "COM812",  # missing-trailing-comma (conflicts with formatter)
    "ISC001",  # single-line-implicit-string-concatenation (conflicts with formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert allowed in tests
    "ARG",     # unused arguments allowed in tests (fixtures)
    "PLR2004", # magic values allowed in tests
]
"__init__.py" = [
    "F401",    # unused imports allowed in __init__.py
]

[tool.ruff.lint.isort]
known-first-party = ["hc_hermes"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = "frontmatter.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=hc_hermes",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
